+++ src/Media/RegenerateWebPMediaCli.php
@@ -12,6 +12,8 @@ namespace EightshiftLibs\Media;
 
 use EightshiftLibs\Cli\AbstractCli;
 use EightshiftLibs\Cli\ParentGroups\CliRun;
+use EightshiftLibs\Helpers\Helpers;
+use Exception;
 use WP_CLI;
 use WP_Query;
 
@@ -20,11 +22,6 @@ use WP_Query;
  */
 class RegenerateWebPMediaCli extends AbstractCli
 {
-	/**
-	 * Media WebP Trait.
-	 */
-	use MediaWebPTrait;
-
 	/**
 	 * Get WP-CLI command parent name
 	 *
@@ -53,10 +50,9 @@ class RegenerateWebPMediaCli extends AbstractCli
 	public function getDefaultArgs(): array
 	{
 		return [
-			'action' => 'generate',
 			'quality' => '80',
 			'ids' => '',
-			'force' => 'false',
+			'allowed_ext' => 'jpg,jpeg,png,bmp',
 		];
 	}
 
@@ -72,21 +68,17 @@ class RegenerateWebPMediaCli extends AbstractCli
 			'synopsis' => [
 				[
 					'type' => 'assoc',
-					'name' => 'action',
-					'description' => 'Action to use "generate" or "delete". Default: generate',
-					'optional' => true,
-					'default' => $this->getDefaultArg('action'),
-					'options' => [
-						'generate',
-						'delete',
-					],
-				],
-				[
-					'type' => 'assoc',
 					'name' => 'quality',
 					'description' => 'Quality of conversion 0-100. Default: 80',
 					'optional' => true,
 					'default' => $this->getDefaultArg('quality'),
+				],
+				[
+					'type' => 'assoc',
+					'name' => 'allowed_ext',
+					'description' => 'Regenerate all supported media to WebP format.',
+					'optional' => true,
+					'default' => $this->getDefaultArg('allowed_ext'),
 				],
 				[
 					'type' => 'assoc',
@@ -97,15 +89,11 @@ class RegenerateWebPMediaCli extends AbstractCli
 				],
 				[
 					'type' => 'assoc',
-					'name' => 'force',
-					'description' => 'Force generation no matter if the file exists. Default: false',
+					'name' => 'only_update_db',
+					'description' => 'Only update the database, not the media files as assumed that the media files are already converted and located on S3 or other storage.',
 					'optional' => true,
-					'default' => $this->getDefaultArg('force'),
-					'options' => [
-						'true',
-						'false',
-					],
-				],
+					'default' => $this->getDefaultArg('skip_skipped'),
+				]
 			],
 			'longdesc' => $this->prepareLongDesc("
 				## USAGE
@@ -122,20 +110,11 @@ class RegenerateWebPMediaCli extends AbstractCli
 				# Regenerate multiple attachments by IDs.
 				$ wp {$this->commandParentName} {$this->getCommandParentName()} {$this->getCommandName()} --ids='16911, 1692, 1302'
 
-				# Force regenerate attachments no matter if they all-ready exist.
-				$ wp {$this->commandParentName} {$this->getCommandParentName()} {$this->getCommandName()} --force='true'
-
 				# Regenerate media with different quality.
 				$ wp {$this->commandParentName} {$this->getCommandParentName()} {$this->getCommandName()} --quality='90'
 
-				# Delete all WebP media formats.
-				$ wp {$this->commandParentName} {$this->getCommandParentName()} {$this->getCommandName()} --action='delete'
-
-				# Delete only one WebP attachment by ID.
-				$ wp {$this->commandParentName} {$this->getCommandParentName()} {$this->getCommandName()} --ids='16911' --action='delete'
-
-				# Delete multiple WebP attachments by ID.
-				$ wp {$this->commandParentName} {$this->getCommandParentName()} {$this->getCommandName()} --ids='16911, 1692, 1302' --action='delete'
+				# Only update the database, not the media files as assumed that the media files are already converted and located on S3 or other storage.
+				$ wp {$this->commandParentName} {$this->getCommandParentName()} {$this->getCommandName()} --only_update_db='true'
 			"),
 		];
 	}
@@ -148,40 +127,53 @@ class RegenerateWebPMediaCli extends AbstractCli
 		$this->getIntroText($assocArgs);
 
 		$quality = $this->getArg($assocArgs, 'quality');
-		$action = $this->getArg($assocArgs, 'action');
 		$ids = $this->getArg($assocArgs, 'ids');
-		$force = $this->getArg($assocArgs, 'force');
+		$allowedExt = $this->getArg($assocArgs, 'allowed_ext');
+		$onlyUpdateDb = (bool) $this->getArg($assocArgs, 'only_update_db');
 
 		$args = [];
 
 		if ($ids) {
 			$args['post__in'] = \explode(',', $ids);
+		} else {
+			$args['post_mime_type'] = \array_map(
+				static function ($item) {
+					return "image/{$item}";
+				},
+				\explode(',', \str_replace(' ', '', $allowedExt))
+			);
 		}
 
 		$options = [
 			'quality' => (int) $quality,
-			'force' => (bool) $force,
 		];
 
-		switch ($action) {
-			case 'delete':
-				$this->deleteMedia($args);
-				break;
-			default:
-				$this->generateMedia($options, $args);
-				break;
+		WP_CLI::log("Depending on the number of attachments, this process may take a while...");
+
+		if ($onlyUpdateDb) {
+			WP_CLI::log("Only updating the database, not the media files as assumed that the media files are already converted and located on S3 or other storage.");
 		}
+
+		WP_CLI::confirm("Are you sure you want to regenerate the media, this process is irreversible and will regenerate all the media files and update the database? Make sure you have a backup of your database!", $assocArgs);
+
+		$this->generateMedia($options, $args, $onlyUpdateDb);
 	}
 
 	/**
-	 * Get Query args
+	 * Generate media.
 	 *
-	 * @param array<string> $args Args to merge on the original array.
+	 * @param array<mixed> $options Options from WP-CLI.
+	 * @param array<mixed> $args Parameters from WP-CLI.
+	 * @param bool $onlyUpdateDb Only update the database, not the media files as assumed that the media files are already converted and locaded on S3 or other storage.
 	 *
-	 * @return array<string>
+	 * @return void
 	 */
-	private function getQueryArgs(array $args): array
+	private function generateMedia(array $options, array $args = [], bool $onlyUpdateDb = false): void
 	{
+		global $wpdb;
+
+		$quality = $options['quality'];
+
 		$defaultArgs = [
 			'post_type' => 'attachment',
 			'post_status' => 'inherit',
@@ -191,108 +183,85 @@ class RegenerateWebPMediaCli extends AbstractCli
 			'update_post_meta_cache' => false,
 			'update_post_term_cache' => false,
 			'fields' => 'ids',
-			'post_mime_type' => \array_map(
-				static function ($item) {
-					return "image/{$item}";
-				},
-				AbstractMedia::WEBP_ALLOWED_EXT
-			),
 		];
 
-		return \array_merge($defaultArgs, $args);
-	}
-
-	/**
-	 * Generate media.
-	 *
-	 * @param array<string, mixed> $options Options from WP-CLI.
-	 * @param array<string> $args Parameters from WP-CLI.
-	 *
-	 * @return void
-	 */
-	private function generateMedia(array $options, array $args = []): void
-	{
-		$quality = $options['quality'];
-		$force = $options['force'];
-
-		$theQuery = new WP_Query($this->getQueryArgs($args));
+		$theQuery = new WP_Query(\array_merge($defaultArgs, $args));
 
 		if (!$theQuery->posts) {
-			WP_CLI::error("No attachments found!");
+			WP_CLI::error("No attachments found with allowed extensions!");
 		}
 
-		foreach ($theQuery->posts as $id) {
-			$title = \get_the_title($id);
+		$skipped = [];
 
-			$original = $this->generateWebPMediaOriginal($id, $quality, $force);
+		foreach ($theQuery->posts as $attachmentId) {
+			$title = \get_the_title($attachmentId);
 
-			WP_CLI::log("Attachment '{$title}' conversion to WebP status: {$id}");
+			WP_CLI::log('--------------------------------------------------');
 
-			if ($original) {
-				WP_CLI::success("Attachment original converted!");
-				WP_CLI::log($original);
-			} else {
-				WP_CLI::warning("Attachment original not converted - already exists!");
-			}
+			WP_CLI::log("Attachment: '{$attachmentId} - {$title}' is being processed...");
 
-			$sizes = $this->generateWebPMediaAllSizes($id, $quality, $force);
+			try {
+				$attachmentMetadata = \get_post_meta($attachmentId, '_wp_attachment_metadata', true);
 
-			if ($sizes) {
-				foreach ($sizes as $size => $sizeValue) {
-					WP_CLI::success("Attachment size {$size} converted!");
-					WP_CLI::log($sizeValue);
+				$mainAttachment = Helpers::convertMediaToWebPById($attachmentId, $quality, $onlyUpdateDb);
+
+				if ($mainAttachment['originalUrl'] !== $mainAttachment['newUrl']) {
+					// phpcs:disable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching
+					$wpdb->update(
+						$wpdb->posts,
+						[
+							'guid' => $mainAttachment['newUrl'],
+							'post_mime_type' => $mainAttachment['newType'],
+						],
+						['ID' => $attachmentId]
+					);
+					// phpcs:enable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching
+
+					WP_CLI::runcommand("search-replace '{$mainAttachment['originalUrl']}' '{$mainAttachment['newUrl']}'");
+
+					$attachmentMetadata['file'] = $mainAttachment['dirnameRelative'] . '/' . $mainAttachment['newFileName'];
 				}
-			} else {
-				WP_CLI::warning("Attachment sizes not converted - already exists or media is to small for additional sizes!");
+
+				$outputSizesMeta = [];
+				if ($sizes = \wp_get_attachment_metadata($attachmentId)['sizes'] ?? []) {
+					foreach ($sizes as $sizeName => $sizeData) {
+						$sizeFullPath = $mainAttachment['dirname'] . '/' . $sizeData['file'];
+						$sizeAttachment = Helpers::convertMediaToWebPByPath($sizeFullPath, $quality, $onlyUpdateDb);
+
+						$sizeData['file'] = $sizeAttachment['newFileName'];
+						$sizeData['mime-type'] = $sizeAttachment['newType'];
+
+						$outputSizesMeta[$sizeName] = $sizeData;
+					}
+				}
+
+				if ($outputSizesMeta) {
+					$attachmentMetadata['sizes'] = $outputSizesMeta;
+				}
+
+				\update_post_meta($attachmentId, '_wp_attachment_metadata', $attachmentMetadata);
+				\update_post_meta($attachmentId, '_wp_attached_file', $mainAttachment['dirnameRelative'] . '/' . $mainAttachment['newFileName']);
+
+				WP_CLI::success("Attachment: '{$attachmentId} - {$title}' converted to WebP format with new URL: {$mainAttachment['newUrl']}");
+			} catch (Exception $e) {
+				WP_CLI::warning("Attachment: '{$attachmentId}' failed with error: '{$e->getMessage()}'! Skipped.");
+				$skipped[$attachmentId] = [
+					'id' => $attachmentId,
+					'title' => $title,
+					'error' => $e->getMessage(),
+				];
+				continue;
 			}
 
 			WP_CLI::log('--------------------------------------------------');
 		}
 
-		\wp_reset_postdata();
-	}
-
-	/**
-	 * Delete media.
-	 *
-	 * @param array<string> $args Parameters from WP-CLI.
-	 *
-	 * @return void
-	 */
-	private function deleteMedia(array $args = []): void
-	{
-		$theQuery = new WP_Query($this->getQueryArgs($args));
-
-		if (!$theQuery->posts) {
-			WP_CLI::error("No attachments found!");
-		}
-
-		foreach ($theQuery->posts as $id) {
-			$title = \get_the_title($id);
-
-			$original = $this->deleteWebPMediaOriginal($id);
-
-			WP_CLI::log("Attachment '{$title}' deleting WebP status: {$id}");
-
-			if ($original) {
-				WP_CLI::success("Attachment original deleted!");
-				WP_CLI::log($original);
-			} else {
-				WP_CLI::warning("Attachment original not deleted - already deleted or missing!");
-			}
-
-			$sizes = $this->deleteWebPMediaAllSizes($id);
-
-			if ($sizes) {
-				foreach ($sizes as $size => $sizeValue) {
-					WP_CLI::success("Attachment size {$size} deleted!");
-					WP_CLI::log($sizeValue);
-				}
-			} else {
-				WP_CLI::warning("Attachment sizes not deleted - already deleted or missing!");
-			}
-
+		if ($skipped) {
 			WP_CLI::log('--------------------------------------------------');
+			WP_CLI::log('Here is the list of skipped attachments:');
+			WP_CLI::log('Skipped attachments count: ' . \count($skipped));
+			$skippedCount = \wp_json_encode($skipped);
+			WP_CLI::log($skippedCount);
 		}
 
 		\wp_reset_postdata();