+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [11.0.0]
+
+### Removed
+
+- `WebPMediaColumnCli`, `WebPMediaColumnExample`, `MediaWebPTrait`, `UseWebPMediaCli` as it is not used anymore.
+
+### Changed
+
+- WebP helpers are now using totally different approach with different hooks and methods. We are no longer creating a new image on upload and using the helper method to detect the extension but we are creating a new image on upload and using using native WordPress functions.
+
+### Updated
+
+- composer packages.
+
 ## [10.12.1]
 
 ### Added
@@ -1034,6 +1048,8 @@ Init setup
 - Gutenberg Blocks Registration.
 - Assets Manifest data.
 
+[12.0.0]: https://github.com/infinum/eightshift-libs/compare/10.12.1...11.0.0
+[11.0.0]: https://github.com/infinum/eightshift-libs/compare/10.12.1...11.0.0
 [10.12.1]: https://github.com/infinum/eightshift-libs/compare/10.12.0...10.12.1
 [10.12.0]: https://github.com/infinum/eightshift-libs/compare/10.11.2...10.12.0
 [10.11.2]: https://github.com/infinum/eightshift-libs/compare/10.11.1...10.11.2